correct PIF RAM SH edge cases

+++ src/cpu/mips_instructions.c
@@ -373,7 +373,7 @@ MIPS_INSTR(mips_sh) {
     s16 offset  = instruction.i.immediate;
     u64 address = get_register(instruction.i.rs);
     address += offset;
-    u16 value = get_register(instruction.i.rt);
+    u32 value = get_register(instruction.i.rt); // A larger value is needed in some cases due to bus weirdness
     u32 physical;
     if (!resolve_virtual_address(address, BUS_STORE, &physical)) {
         on_tlb_exception(address);
